MDL-78616 bigbluebuttonbn: Added new meeting timing features
Added ability to hide presentation file from viewers. This is useful if the file is used during the meeting for a white board, but contains no other information.
Added ability to allow moderators to join before the meeting opening time. This would give time for the moderators to join and get ready before the viewers join.
Added ability to schedule a meeting start time that occurs after the meeting opening time. (Impacts calendar event scheduling) The calendar event will be scheduled for the starting time instead of the opening time. This allows meeting participants to join the meeting earlier than the scheduled start time to mimic arriving at a physical room before the start of a meeting or class.
Added appropriate settings to disable and hide these new features if not wanted.
Set defaults to not change any behaviors when plugin is upgraded.